requirements.txt
asgiref==3.7.2
cffi==1.15.1
channels==4.0.0
Django==4.2.5
django-crispy-forms==2.0
factory-boy==3.3.0
Faker==19.3.1
misaka==2.1.1
Pillow==10.0.0
pycparser==2.21
pypng==0.20220715.0
python-dateutil==2.8.2
qrcode==7.4.2
reportlab==4.0.4
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1